More inv t LUT to class
