+++ TrackletAlgorithm/TrackletCalculator_calculate_LXD1.h
@@ -2,6 +2,7 @@
 #include "ap_int.h"
 #include "Constants.h"
 #include "TrackletProjectionCalculator.h"
+#include "TrackletLUTs.h"
 
 
 static void init_idr_overlap(ap_uint<18> *lut) {
@@ -143,7 +144,10 @@ void TC::calculate_LXD1 (
   *der_phiL_output = -(irinv_new >> (1+3));
   *der_zL_output = it_new >> 3;
 
-  ap_uint<20> itinv = lut_itinv(abs(it_new)&4095);
+  static const InvtLUT lut_itinv;
+
+  //ap_uint<20> itinv = lut_itinv(abs(it_new)&4095);
+  ap_uint<20> itinv = lut_itinv.lookup(abs(it_new)&4095);
 
   constexpr int izproj0 = zmean[projectionDisks[Seed][0]];
   constexpr int izproj1 = zmean[projectionDisks[Seed][1]];
